autotimer: add autotimer id to xml if requested from webif
please note that the ids are not what one would call persistent as they
are assigned as the configuration is read by a counter. however, since
the order inside the configuration does not change without human
interaction and will be the same again on subsequent reads. to make it
short: don't use the id you read some days ago, but rather double check it
before using it.