autotimer: reduce xml config version to 5 again
this version number only indicates the version of the config file and
only intended for internal use by the parser. increasing the number
should only be done if there is an incompatible change in syntax/meaning
of values to the previous version(s), which is why the "fallback
algorithm" parseConfigOld is called. it is to be modified according to
the changes in order for bumping the version to make any sense.

this might lead to the same trouble that people could have run into when
installing this version in the first place (config COULD be read
incorrectly) but let's hope there were no errors in the first place.