autotimer: pick iso8859-15 as default encoding with german gui
please note that this changes the previous default behavior and might
lead to unexpected results, as the default encoding is not saved. so
switching languages might result in the encoding to change. this does
however not affect timers already in memory.
in general this should improve general user experience while some might
be confused if they change the on-screen language a lot.