AutoTimer.py: workaround twisted deadlock
if the autotimer was running during shutdown twisted would lock up (the
child thread was waiting indefinitely).
this works around this by giving the main thread operation 30s to finish
which should be plenty in our use case.
if someone finds a way to reliably detect if twisted is currently trying
to shut down please fix the crash condition :)