AutoTimer.py: smaller cleanup and threading optimization
- factor out loop previously done with each enabled timer into another
  function to improve readbility a bit (still not the easiest piece of
  code to read if you have no idea how the plugin works)
- change an embedded into an instance method
- rename some variables as comments suggested, follow general naming
  scheme with others

the threading change is the most interesting in regards to the most
recent changes.
instead of many context changes we now only switch once for each timer.
this increases the chances to block the main thread long enough for the
user to notice, but this way it's still better than before threading and
we should finish in under 10 minutes again :P

please note that while i tested this code i'd consider it experimental
for some days.